requirements.txt
altair==4.2.2
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
entrypoints==0.4
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.0
gitdb==4.0.11
GitPython==3.1.43
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
ipykernel==6.29.5
ipython==8.27.0
itables==2.2.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Levenshtein==0.26.0
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multimethod==1.12
mysql-connector-python==9.1.0
mysqlclient==2.2.4
narwhals==1.8.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyWavelets==1.7.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.3
streamlit==1.38.0
streamlit-aggrid==1.0.5
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.6
watchdog==4.0.2
wcwidth==0.2.13
wordcloud==1.9.3
ydata-profiling==4.10.0
